requirements.txt
# PyTorch & verwandte Pakete
torch>=1.10.0
torchvision>=0.11.0
torchaudio>=0.10.0

# Numerische Berechnungen & Datenverarbeitung
numpy>=1.21.2
scipy>=1.7.1
pandas>=1.3.3
pyyaml>=5.4.1

# NLP & Textverarbeitung
nltk>=3.6.5
spacy>=3.2.0
transformers>=4.15.0
sentencepiece>=0.1.96

# Evaluierung & Visualisierung
scikit-learn>=0.24.2
matplotlib>=3.4.3
seaborn>=0.11.2

# Web-API
flask>=2.0.2
flask-restful>=0.3.9

# Logging & Debugging
loguru>=0.5.3

# Fortschrittsanzeige während des Trainings
tqdm>=4.62.3

# Tests und Qualitätssicherung
pytest>=6.2.5
pytest-cov>=2.12.1
